chore: move JsonConverters into files matching the respective class names (#558)
